requirements.txt
langchain~=0.2.14
langchain-community
langchain-openai
openai
pillow
chainlit~=1.1.402
python-dotenv~=1.0.1
aiohttp~=3.10.3